#41 add 2 more parameters in useEffect and separate createContext in a separate file with modifications to the function callers
