+++ Project/Frontend/src/AllOrders.jsx
@@ -3,7 +3,7 @@ import { Box, Typography, List, ListItem, ListItemText, Divider, Grid, Paper, Bu
 import ExpandLessIcon from '@mui/icons-material/ExpandLess';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import RefreshIcon from '@mui/icons-material/Refresh';
-import { CartContext } from './CartContext';
+import { CartContext } from './CartContext/CartContextContext.jsx';
 
 const AllOrders = () => {
     // State to store orders and the currently expanded order ID
