+++ Project/Frontend/src/CartContext/CartContextContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const CartContext = createContext();
